Create model of JWL DB & implement import logic
The backup file of the JW Library is essentially
a zip file containing a manifest and a SQLite database.
My idea is to parse this database and create a
in-memory representation, so its easier and faster
to work on the data.

This commit adds the model of the database and adds logic to
import a JW Library backup file into the local Database model.
So calling `ImportJWLBackup` on a Database{} struct should populate
it.