Include empty user_data.db
This empty (so only containing the
base-structure) SQLite DB allows to add all data
after merging backups internally. It is not
possible to simply create a new SQLite DB on
the fly and adding the JW Library schema, as
this won't include data of the index etc. I've
tried it and it resulted in a

> INSERT INTO Tag (Type, Name) VALUES (0, 'Favorite'); Unknown error calling sqlite3_step (19: UNIQUE constraint failed: Tag.Type, Tag.Name)