:recycle: Use slice of pointers
:recycle: By using pointers to the structs instead
of the structs themself, we are able to
easily update the structs directly when
merging entries. This should also give
us a performance boost, as we need to
copy less data back-and-forth, especially
if there are a lot of entries.

:bulb: Also added some more comments
to clarify how some functionality is
working. Especially when using reflection
it can become quite confusing otherwise.