:art: Extend Model & make it public
:recycle: Make Model public, so we can use it in the
the merge package.

:art: Adding functions `SetID`, `UniqueKey`, and
`Equals` to the Model interface. These will be helpful when
building a general merge function:

* `SetID`: Allows to set the ID of a entry after
merging, should the ID have changed in the process.
* `UniqueKey`:  We can use the uniqueKey to
distinguish different entries of the same type and
find conflicts that would need to be solved.
* `Equals`: Makes it possible to quickly compare
to entries and see if their content (so ID
not included) is equal.