:sparkles: General merge implementation
The `merge` function allows to merge two slices
of structs that implement the `Model`-interface.
If it detects a collision it returns an error, so the
caller is able to handle the collision himself. This
could mean adding functionality in-code or asking
a user for directions.

`tryMergeWithConflictSolver` will use `merge`
and a given conflictSolver to try to automatically
solve mergeConflicts if possible. If it is not able
to decrease the mergeConflicts anymore, it forwards
the conflicts to the caller.

As the results of `merge`, `prepareMergeSolution`,
and `tryMergeWithConflictSolver`
are supposed to be used by upper-layer functions
(which should have good-enough tests) and as they
consume the general `Model` interface, I'm leaving
these functions untested.