Merge Locations
Locations indicate where a Bookmark, Note, Mark,
etc. is located and is not directly user controlled.
If two Locations collide, the only difference may
be the title (in that one doesn't have one). Therefore
we can safely use a function to solve a merge
conflict, which essentially chooses the Location
which has a title.