:sparkles: Merge UserMark & BlockRange
Merging UserMarks and BlockRanges was not possible
using the generic merge function. It needs additional
functionalities to detect overlapping
BlockRanges/markings.

UserMarks & BlockRanges are first joined, so we are able to combine
them by Location (e.g. a page in a book) and later by Identifier (e.g.
a paragraph of a page or a verse in the bible). After that
BlockRanges are sorted by StartToken in increasing order
to detect overlapping intervals. Collisions because of
duplicates are handled automatically, for others the caller
needs to decide which side should be chosen. It is also
possible to combine them by tweaking the solution.