:bug: DatabaseName in manifest was ignored
When importing a backup, the DatabaseName was
hard-coded to `user_data.db`, which would not work
for backups exported by a JWL running on Windows
(there the file is called `userData.db`.

This commit changes the handling of manifests,
adding a proper import/export functionality,
splitting it out of the validate method. This allows us
to use the information within the manifest for other
use cases (like getting the propere DatabaseName).