:bug: Don't overwrite original entry while merging
Previously, calling a merge function resulted in the
original database entries being tweaked (mainly their
IDs), as only their pointers were copied. This
commit changes it to create copies of the entries
and add them to the merged result, so the copied
entries can be safely tweaked. This allows to
re-use the same Database for multiple merge
attempts.