:bug: Combine equal TagMap with different position
Because the position of a TagMap was included
in the uniqueKey, the same TagMap with a different
position would be considered as a different TagMap
entry, resulting in a possible

> UNIQUE constraint failed: TagMap.TagId, TagMap.NoteId

error.

This removes the position from the uniqueKey of
TagMaps and adds tests to detect the problem
in the future.