:sparkles: Mobile wrapper using Gomobile
This introduces the package `go-jwlm/gomobile` that enables
integrating go-jwlm into a mobile app. Because of current
limitations in Gomobile
(see
https://godoc.org/golang.org/x/mobile/cmd/gobind#hdr-Type_restrictions),
the structs and most functionalities had to be wrapped.

The basic structure looks like this:
* The `DatabaseWrapper` wraps the left, right, and merged
databases. It manages importing and exporting
* The merge functions (e.g. MergeLocations etc.) are called
on the DatabaseWrapper
* Conflicts are handled using the `MergeConflictsWrapper`, which
exports methods to view and solve possible conflicts
* To get some rough statistics, the `DatabaseStats` struct can be
created

To create the actual binding, Gomobile needs to be installed
(see https://github.com/golang/go/wiki/Mobile). Change into the
go-jwlm/gomobile folder and run `gomobile bind -target ios` (in this
case to target iOS), which will create the binding that can later
be imported. The (WIP) implementation for iOS can be found here:
https://github.com/AndreasSko/ios-jwlm