:recycle: Use struct for relatedEntries
Instead of using a slice of relatedEntries, it is better
to set them specifially. This allows to easily detect
if there actually exists, for example, a related
Location, instead of needing to go through the
whole slice.

This change introduces the model.Related struct,
which has the necessary fields to represent
all possible related entries.