:bug: Don't count similar collisions multiple times
If on one side multiple BlockRanges collide with the
same BlockRange on the other side, the algorithm
shouldn't count all collisions, as further
possible collisions depend on the outcome of the
first resolution: If the first solution discards the
one entry that overlapped with the other ones, all
further collisions have been solved with it.

This changes it so all other possible collisions are
skipped first. They are either handled automatically
or will be detected in the next run of the function.