:bug: Make copy for merged UMBRs
The `UserMarkBlockRange` entries that went
into a merged solution pointed to the same ones
returned from a previous mergeConflict and so
also went into possible mergeSolutions. As
`mergeUMBR` updates the userMarkIDs after
a successful merge, the IDs in the
mergeSolutions were updated too (as they were
in fact the SAME). This resulted in out-of-sync
IDs when running a merge again with the
previous mergeSolution.

This change simply creates a copy before
changing the IDs. Also added a test to detect
regressions in the future.