:bug: Consider partially migrated Bibles
When the Study Edition of the Revised New
World Translation was published, JW Library
offered to migrate all previous markings from
the Standard Edition to the Study Edition. It
did that by simply replacing the KeySymbols
in Locations from `nwt` to `nwtsty`, while
keeping the UserMarkGUIDs etc. the same.

When merging two backups with the same
markings, while one side has already been
migrated to `nwtsty`, markings can't be
detected as duplicate or overlapping
anymore: They techically belong to different
locations, while their UserMarkGUID stays the
same. This results in UserMarks with the
same GUID, but different locations (one for
`nwt` and one for `nwtsty`), which, when
exporting, leads to a unique constraint
violation.

This commits adds a check that checks if one
side has migrated to the Study Edition, while
the other one has not. In this case, it will
also migrate the other side to `nwtsty`.