Gomobile: Download catalog & lookup publications
Gomobile wraps the functions of the
`publication` package, so we are able to use
them in mobile devices. This also has the nice
effect, that we only need to care about one
functionality for downloading the catalog
instead of making it platform specific.

When a download of the catalog is initalized
using `DownloadCatalog`, it returns a
`DownloadManager`, which holds all
necessary information to keep track of the
download and even cancel it mid-flight.