:zap: More efficient regex for replaceUMBRConflictsWithSolution
:zap: To get the number before the UniqueKey
of the UMBR conflictSolution, we used
`ReplaceAllString`, which does more work
than needed (we don't need to
replace the string, but just get the number).
This changes it to use the proper
`FindStringSubmatch`, which sped up the
function significantly:

Before, a merge of two 2.5MB backups took
a second within `replaceUMBRConflictsWithSolution`,
now it's down to 0.05s!

:recycle: To make the sorting more testable, I split it
into its own function and added proper test cases for it.