Decide nwtsty migration on collision, not via heuristic
The current approach of determining the need
of migrating
from `nwt` to `nwtsty` (introduced with !46) is
using a heuristic, which basically checks if one side
has significantly more `nwtsty` entries for a specific
language than the other side.
This heuristic worked quite well, but can fail if the
difference between both sides is not high enough.
This can lead to collisions of UserMarkGUIDs.

This introduces a new approach, which will check
both sides for the same UserMarkGUID. If it finds a
common GUID, it will check if the location of one
UserMark is in `nwt`, while the other one is in `nwstsy`.
If this is the case, it indicates the need for
migrating to `nwtsty`, as obviously one side has
already been migrated.

The logic for the migration is now bundled in kind of a
"PreMergeHook", which bundles necessary functionality
for preparing the databases for merging.