Cleanup userMark duplicates after nwtsty migration
The migration from `nwt` to `nwtsty` is working
properly for "simple" locations, e.g. bible chapters.
However, the locations for appendices have
DocumentIDs, which are not easily migratable
(we would need to find the documentID
corresponding to the `nwstsy`). Therefore, we skip
those locations.

Skipping the locations, however, introduces the
risk of UserMark collisions again, as the merger
is using the locationIDs for finding those. As this
is a very limited edge-case, I decided to simply
remove the duplicate from the old `nwt` side
completely. This shouldn't be a problem, as
markings on the `nwstsy` can be considered
newer.