Skip entries that don't match schema on import
For some reason, it's possible that a table contains NULL
entries, even if the schema doesn't allow for that. Currently,
an import would then fail with a `Scan error`. As the entry anyway
is not valid, we can also simply skip it and continue with the
import as usual.

This PR adds a `isNullableMismatch` function, which can check
if a given error is due to a mismatch of a NULL entry and a
NOT-NULL schema. If we detect this kind of issue while scanning
a table, we don't need to fail, but can simply skip to the next
entry.

Fixes #115