:bug: Skip UserMarks with no BlockRange attached
The "duplicate" check for UserMarks is based on the collision check of
the BlockRanges. So if no collision is detected for a duplicate marking,
the duplicate itself will not be detected. This is an issue for
UserMarks that don't have any associated BlockRanges: Duplicates won't
be detected, which in the end results in a
`UNIQUE constraint failed: UserMark.UserMarkGuid` error when exporting.

This adds a check for markings that don't have any BlockRanges attached
to them. They will be removed now, as they aren't useful anyway.