Cancel import of playlists exists, with option to skip
The new JW Library version 14 now supports
playlists. Implementing the merging of them
will take more time and testing to ensure
everything works as expected. Until then
we need to make sure that we don't
accidentally attempt to merge playlists, which
might break the whole backup.

With this commit we are adding a check
that will make sure that all playlists related
tables are empty. Also, as playlists and their
entries are implemented via tags, it checks
for any playlist entries there.
It's possible to skip the playlists check via
the `SkipPlaylists` flag, that has been
added to the `Database` struct. If it is set,
playlist entries will be removed while
importing, ensuring that the rest of the
data can be merged without any issue.