Support schema version range
For some reason the latest JW Library version
moved the schema version back to 13.
In the hope that in a next version schema 14
still doesn't contain changes, we are now
allowing a range of supported schema versions.
This means a backup can be of version 13
AND 14 and still be considered compatible.