Revert Schema version 14 support
It seems that JW Library has rolled back
schema version 14 and exported backups
now use v13 again. As mentioned in
https://github.com/AndreasSko/go-jwlm/pull/147#issuecomment-1678741019
it seems they change something in the
data between those two version, so it's not
possible to support both schema versions
at the same time. Therefore, we are reverting
back to version 13.