:bug: InputFields merged to wrong publication
Since the last version of JW Library, a location
can have an empty MepsLanguage. Probably
this has been implemented so that personal
notes in a publication will be shown regardless
of its language (i.e. a ticked box is ticked
both in english as well as in any other language
of the same publication).
The model for Location, however, wasn't
updated accordingly when adding support
for it in https://github.com/AndreasSko/go-jwlm/pull/141.
This meant that newly created entries
in input fields might be skipped because
of a `NullableMismatch`, in the end resulting
in inputField entries showing up at wrong
locations or being missing completely.

This PR fixes that issue now.