Support custom temporary directory
Users on iOS reported issues when
importing backups:

> Error while creating temporary directory: mkdir /var/folders/t_/0w12llq155s3bcljc16_hj_c0000gn/T/go-jwlm1513177742: operation not permitted

It looks like Go might now always
be able to determine the correct
path for temporary directories.

With this PR we now support
customizing the tempDir. This will
be used in the iOS app to create
a tempDir via Swift and then providing
it to go-jwlm.